add firmware recipes, based on linux-firmware
The default linux-firmware recipe adds 80MB of irrelevant blobs to
the system image and breaks the bitbake dependency resolution so that
it doesn't even add the specific sub-package you ask it too.

This recipe removes all the sub-packages from linux-firmware, adds a
sub-package for the Realtek 8168 firmware and will ONLY add the
sub-package you ask for.